requirements.txt
google-adk>=0.5.0
python-dotenv>=1.0.0
sarvamai>=0.1.0
google-generativeai>=0.3.1
requests>=2.31.0
openai>=1.0.0 